feat(push): send fast-forward merge status event
when a proposal was merged using fast-forward status rather than
by creating a three way merge commit.

if there are multiple revisions, the first one that contains merged
proposal tip will be referenced.

if there are multiple proposals that contain one of the commits,
they will all be marked as merged.

the nip needs to be updated as there is no single `merge-commit-id`
so that tag needs to support multiple values